+++ source/RRTagController.swift
@@ -14,60 +14,59 @@ struct Tag {
     var textContent: String
 }
 
-let colorUnselectedTag = UIColor.whiteColor()
+let colorUnselectedTag = UIColor.white
 let colorSelectedTag = UIColor(red:0.22, green:0.7, blue:0.99, alpha:1)
 
 let colorTextUnSelectedTag = UIColor(red:0.33, green:0.33, blue:0.35, alpha:1)
-let colorTextSelectedTag = UIColor.whiteColor()
+let colorTextSelectedTag = UIColor.white
 
 class RRTagController: UIViewController, UICollectionViewDelegate, UICollectionViewDataSource, UICollectionViewDelegateFlowLayout {
 
-    private var tags: Array<Tag>!
-    private var navigationBarItem: UINavigationItem!
-    private var leftButton: UIBarButtonItem!
-    private var rigthButton: UIBarButtonItem!
-    private var _totalTagsSelected = 0
-    private let addTagView = RRAddTagView(frame: CGRectMake(0, 0, UIScreen.mainScreen().bounds.size.width, 64))
-    private var heightKeyboard: CGFloat = 0
+    fileprivate var tags = [Tag]()
+    fileprivate var navigationBarItem: UINavigationItem!
+    fileprivate var leftButton: UIBarButtonItem!
+    fileprivate var rigthButton: UIBarButtonItem!
+    fileprivate var _totalTagsSelected = 0
+    fileprivate let addTagView = RRAddTagView(frame: CGRect(x: 0, y: 0, width: UIScreen.main.bounds.size.width, height: 64))
+    fileprivate var heightKeyboard: CGFloat = 0
     
-    var blockFinih: ((selectedTags: Array<Tag>, unSelectedTags: Array<Tag>) -> ())!
-    var blockCancel: (() -> ())!
+    var blockFinish: ((_ selectedTags: [Tag], _ unSelectedTags: [Tag]) -> Void)!
+    var blockCancel: (() -> Void)!
 
     var totalTagsSelected: Int {
         get {
-            return self._totalTagsSelected
+            return _totalTagsSelected
         }
         set {
-            if newValue == 0 {
-                self._totalTagsSelected = 0
+            guard newValue > 0 else {
+                _totalTagsSelected = 0
                 return
             }
-            self._totalTagsSelected += newValue
-            self._totalTagsSelected = (self._totalTagsSelected < 0) ? 0 : self._totalTagsSelected
-            self.navigationBarItem = UINavigationItem(title: "Tags")
-            self.navigationBarItem.leftBarButtonItem = self.leftButton
-            if (self._totalTagsSelected == 0) {
-                self.navigationBarItem.rightBarButtonItem = nil
+            _totalTagsSelected += newValue
+            _totalTagsSelected = _totalTagsSelected < 0 ? 0 : _totalTagsSelected
+            navigationBarItem = UINavigationItem(title: "Tags")
+            navigationBarItem.leftBarButtonItem = leftButton
+            if (_totalTagsSelected == 0) {
+                navigationBarItem.rightBarButtonItem = nil
+            } else {
+                navigationBarItem.rightBarButtonItem = rigthButton
             }
-            else {
-                self.navigationBarItem.rightBarButtonItem = self.rigthButton
-            }
-            self.navigationBar.pushNavigationItem(self.navigationBarItem, animated: false)
+            navigationBar.setItems([navigationBarItem], animated: false)
         }
     }
     
     lazy var collectionTag: UICollectionView = {
         let layoutCollectionView = UICollectionViewFlowLayout()
         layoutCollectionView.sectionInset = UIEdgeInsets(top: 0, left: 20, bottom: 0, right: 20)
-        layoutCollectionView.itemSize = CGSizeMake(90, 20)
+        layoutCollectionView.itemSize = CGSize(width: 90, height: 20)
         layoutCollectionView.minimumLineSpacing = 10
         layoutCollectionView.minimumInteritemSpacing = 5
         let collectionTag = UICollectionView(frame: self.view.frame, collectionViewLayout: layoutCollectionView)
-        collectionTag.contentInset = UIEdgeInsets(top: 84, left: 0, bottom: 20, right: 0)
+        collectionTag.contentInset = UIEdgeInsets(top: 64, left: 0, bottom: 20, right: 0)
         collectionTag.delegate = self
         collectionTag.dataSource = self
-        collectionTag.backgroundColor = UIColor.whiteColor()
-        collectionTag.registerClass(RRTagCollectionViewCell.self, forCellWithReuseIdentifier: RRTagCollectionViewCellIdentifier)
+        collectionTag.backgroundColor = .white
+        collectionTag.register(RRTagCollectionViewCell.self, forCellWithReuseIdentifier: RRTagCollectionViewCellIdentifier)
         return collectionTag
     }()
     
@@ -75,115 +74,110 @@ class RRTagController: UIViewController, UICollectionViewDelegate, UICollectionV
         let addNewTagCell = RRTagCollectionViewCell()
         addNewTagCell.contentView.addSubview(addNewTagCell.textContent)
         addNewTagCell.textContent.text = "+"
-        addNewTagCell.frame.size = CGSizeMake(40, 40)
-        addNewTagCell.backgroundColor = UIColor.grayColor()
+        addNewTagCell.frame.size = CGSize(width: 40, height: 40)
+        addNewTagCell.backgroundColor = UIColor.gray
         return addNewTagCell
     }()
     
     lazy var controlPanelEdition: UIView = {
-        let controlPanel = UIView(frame: CGRectMake(0, UIScreen.mainScreen().bounds.size.height + 50, UIScreen.mainScreen().bounds.size.width, 50))
-        controlPanel.backgroundColor = UIColor.whiteColor()
+        let controlPanel = UIView(frame: CGRect(x: 0, y: UIScreen.main.bounds.size.height + 50, width: UIScreen.main.bounds.size.width, height: 50))
+        controlPanel.backgroundColor = .white
         
-        let buttonCancel = UIButton(frame: CGRectMake(10, 10, 100, 30))
-        buttonCancel.layer.borderColor = UIColor(red:0.88, green:0.88, blue:0.88, alpha:1).CGColor
+        let buttonCancel = UIButton(frame: CGRect(x: 10, y: 10, width: 100, height: 30))
+        buttonCancel.layer.borderColor = UIColor(red:0.88, green:0.88, blue:0.88, alpha:1).cgColor
         buttonCancel.layer.borderWidth = 2
-        buttonCancel.backgroundColor = UIColor.whiteColor()
-        buttonCancel.setTitle("Cancel", forState: UIControlState.Normal)
-        buttonCancel.setTitleColor(UIColor.blackColor(), forState: UIControlState.Normal)
-        buttonCancel.titleLabel?.font = UIFont.boldSystemFontOfSize(17)
+        buttonCancel.backgroundColor = .white
+        buttonCancel.setTitle("Cancel", for: UIControlState())
+        buttonCancel.setTitleColor(.black, for: UIControlState())
+        buttonCancel.titleLabel?.font = .boldSystemFont(ofSize: 17)
         buttonCancel.layer.cornerRadius = 15
-        buttonCancel.addTarget(self, action: "cancelEditTag", forControlEvents: UIControlEvents.TouchUpInside)
+        buttonCancel.addTarget(self, action: #selector(RRTagController.cancelEditTag), for: .touchUpInside)
 
-        let buttonAccept = UIButton(frame: CGRectMake(UIScreen.mainScreen().bounds.size.width - 110, 10, 100, 30))
-        buttonAccept.layer.borderColor = UIColor(red:0.88, green:0.88, blue:0.88, alpha:1).CGColor
+        let buttonAccept = UIButton(frame: CGRect(x: UIScreen.main.bounds.size.width - 110, y: 10, width: 100, height: 30))
+        buttonAccept.layer.borderColor = UIColor(red:0.88, green:0.88, blue:0.88, alpha:1).cgColor
         buttonAccept.layer.borderWidth = 2
-        buttonAccept.backgroundColor = UIColor.whiteColor()
-        buttonAccept.setTitle("Create", forState: UIControlState.Normal)
-        buttonAccept.setTitleColor(UIColor.blackColor(), forState: UIControlState.Normal)
-        buttonAccept.titleLabel?.font = UIFont.boldSystemFontOfSize(17)
+        buttonAccept.backgroundColor = .white
+        buttonAccept.setTitle("Create", for: UIControlState())
+        buttonAccept.setTitleColor(.black, for: UIControlState())
+        buttonAccept.titleLabel?.font = .boldSystemFont(ofSize: 17)
         buttonAccept.layer.cornerRadius = 15
-        buttonAccept.addTarget(self, action: "createNewTag", forControlEvents: UIControlEvents.TouchUpInside)
+        buttonAccept.addTarget(self, action: #selector(RRTagController.createNewTag), for: .touchUpInside)
         
         controlPanel.addSubview(buttonCancel)
         controlPanel.addSubview(buttonAccept)
         return controlPanel
     }()
-    
+
     lazy var navigationBar: UINavigationBar = {
-        let navigationBar = UINavigationBar(frame: CGRectMake(0, 0, UIScreen.mainScreen().bounds.size.width, 64))
+        let navigationBar = UINavigationBar(frame: CGRect(x: 0, y: 0, width: UIScreen.main.bounds.size.width, height: 64))
         
         self.navigationBarItem = UINavigationItem(title: "Tags")
-        self.navigationBarItem.leftBarButtonItem = self.leftButton
-        
-        navigationBar.pushNavigationItem(self.navigationBarItem, animated: true)
+        self.navigationBarItem.leftBarButtonItem = leftButton
+
+        navigationBar.setItems([navigationBarItem], animated: true)
         navigationBar.tintColor = colorSelectedTag
         return navigationBar
     }()
     
-    func cancelTagController() {
-        self.dismissViewControllerAnimated(true, completion: { () -> Void in
+    @objc func cancelTagController() {
+        self.dismiss(animated: true) {
             self.blockCancel()
-        })
+        }
     }
     
-    func finishTagController() {
-        var selected: Array<Tag> = Array()
-        var unSelected: Array<Tag> = Array()
+    @objc func finishTagController() {
+        var selected = [Tag]()
+        var unSelected = [Tag]()
         
         for currentTag in tags {
             if currentTag.isSelected {
                 selected.append(currentTag)
-            }
-            else {
+            } else {
                 unSelected.append(currentTag)
             }
         }
-        self.dismissViewControllerAnimated(true, completion: { () -> Void in
-            self.blockFinih(selectedTags: selected, unSelectedTags: unSelected)
-        })
-    }
-    
-    func cancelEditTag() {
-        self.view.endEditing(true)
-        UIView.animateWithDuration(0.5, delay: 0, usingSpringWithDamping: 0.4,
-            initialSpringVelocity: 0.4, options: UIViewAnimationOptions.allZeros, animations: { () -> Void in
-            self.addTagView.frame.origin.y = 0
-            self.controlPanelEdition.frame.origin.y = UIScreen.mainScreen().bounds.size.height
-            self.collectionTag.alpha = 1
-            }) { (anim:Bool) -> Void in
-            
+        self.dismiss(animated: true) {
+            self.blockFinish(selected, unSelected)
         }
     }
     
-    func createNewTag() {
-        let spaceSet = NSCharacterSet.whitespaceCharacterSet()
-        let contentTag = addTagView.textEdit.text.stringByTrimmingCharactersInSet(spaceSet)
+    @objc func cancelEditTag() {
+        view.endEditing(true)
+        UIView.animate(withDuration: 0.5, delay: 0, usingSpringWithDamping: 0.4, initialSpringVelocity: 0.4, options: [], animations: {
+            self.addTagView.frame.origin.y = self.navigationBar.frame.minY
+            self.addTagView.alpha = 0
+            self.controlPanelEdition.frame.origin.y = UIScreen.main.bounds.size.height
+            self.collectionTag.alpha = 1
+        })
+    }
+    
+    @objc func createNewTag() {
+        let spaceSet = CharacterSet.whitespaces
+        let contentTag = addTagView.textEdit.text.trimmingCharacters(in: spaceSet)
         if strlen(contentTag) > 0 {
             let newTag = Tag(isSelected: false, isLocked: false, textContent: contentTag)
-            tags.insert(newTag, atIndex: tags.count)
+            tags.insert(newTag, at: tags.count)
             collectionTag.reloadData()            
         }
         cancelEditTag()
     }
     
-    func collectionView(collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
         return tags.count + 1
     }
     
-    func collectionView(collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout,
-        sizeForItemAtIndexPath indexPath: NSIndexPath) -> CGSize {
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout,
+        sizeForItemAt indexPath: IndexPath) -> CGSize {
             if indexPath.row < tags.count {
                 return RRTagCollectionViewCell.contentHeight(tags[indexPath.row].textContent)
             }
-            return CGSizeMake(40, 40)
+            return CGSize(width: 40, height: 40)
     }
     
-    func collectionView(collectionView: UICollectionView, didSelectItemAtIndexPath indexPath: NSIndexPath) {
-        let selectedCell: RRTagCollectionViewCell? = collectionView.cellForItemAtIndexPath(indexPath) as? RRTagCollectionViewCell
+    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
+        let selectedCell: RRTagCollectionViewCell? = collectionView.cellForItem(at: indexPath) as? RRTagCollectionViewCell
         
         if indexPath.row < tags.count {
-            var currentTag = tags[indexPath.row]
-            
             if tags[indexPath.row].isSelected == false {
                 tags[indexPath.row].isSelected = true
                 selectedCell?.animateSelection(tags[indexPath.row].isSelected)
@@ -197,39 +191,40 @@ class RRTagController: UIViewController, UICollectionViewDelegate, UICollectionV
         }
         else {
             addTagView.textEdit.text = nil
-            UIView.animateWithDuration(0.5, delay: 0, usingSpringWithDamping: 0.4, initialSpringVelocity: 0.4,
-                options: UIViewAnimationOptions.allZeros, animations: { () -> Void in
+            var margin: CGFloat = 40
+            if #available(iOS 11.0, *), navigationBar.prefersLargeTitles {
+                margin = 0
+            }
+            UIView.animate(withDuration: 0.5, delay: 0, usingSpringWithDamping: 0.4, initialSpringVelocity: 0.4, options: [], animations: {
                 self.collectionTag.alpha = 0.3
-                self.addTagView.frame.origin.y = 64
-                }, completion: { (anim: Bool) -> Void in
-                    self.addTagView.textEdit.becomeFirstResponder()
-                    println("")
+                self.addTagView.frame.origin.y = self.navigationBar.frame.maxY - margin
+                self.addTagView.alpha = 1
+            }, completion: { _ in
+                self.addTagView.textEdit.becomeFirstResponder()
             })
         }
     }
     
-    func collectionView(collectionView: UICollectionView, cellForItemAtIndexPath indexPath: NSIndexPath) -> UICollectionViewCell {
-        let cell: RRTagCollectionViewCell? = collectionView.dequeueReusableCellWithReuseIdentifier(RRTagCollectionViewCellIdentifier, forIndexPath: indexPath) as? RRTagCollectionViewCell
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        let cell: RRTagCollectionViewCell? = collectionView.dequeueReusableCell(withReuseIdentifier: RRTagCollectionViewCellIdentifier, for: indexPath) as? RRTagCollectionViewCell
         
         if indexPath.row < tags.count {
             let currentTag = tags[indexPath.row]
             cell?.initContent(currentTag)
-        }
-        else {
+        } else {
             cell?.initAddButtonContent()
         }
         return cell!
     }
     
-    func keyboardWillShow(notification: NSNotification) {
+    @objc func keyboardWillShow(_ notification: Notification) {
         // TODO: change value
         if let userInfo = notification.userInfo {
-            if let keyboardSize = (userInfo[UIKeyboardFrameBeginUserInfoKey] as? NSValue)?.CGRectValue() {
+            if let keyboardSize = (userInfo[UIKeyboardFrameEndUserInfoKey] as? NSValue)?.cgRectValue {
                 heightKeyboard = keyboardSize.height
-                UIView.animateWithDuration(0.5, delay: 0, usingSpringWithDamping: 0.4, initialSpringVelocity: 0.4,
-                    options: UIViewAnimationOptions.allZeros, animations: { () -> Void in
+                UIView.animate(withDuration: 0.5, delay: 0, usingSpringWithDamping: 0.4, initialSpringVelocity: 0.4, options: [], animations: {
                     self.controlPanelEdition.frame.origin.y = self.view.frame.size.height - self.heightKeyboard - 50
-                }, completion: nil)
+                })
             }
         }
         else {
@@ -237,47 +232,73 @@ class RRTagController: UIViewController, UICollectionViewDelegate, UICollectionV
         }
     }
     
-    func keyboardWillHide(notification: NSNotification) {
+    @objc func keyboardWillHide(_ notification: Notification) {
         heightKeyboard = 0
+    }
+
+    override var prefersStatusBarHidden: Bool {
+        return true
     }
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        self.view.backgroundColor = UIColor.whiteColor()
+        self.view.backgroundColor = UIColor.white
 
-        leftButton = UIBarButtonItem(title: "Cancel", style: UIBarButtonItemStyle.Done, target: self, action: "cancelTagController")
-        rigthButton = UIBarButtonItem(title: "OK", style: UIBarButtonItemStyle.Done, target: self, action: "finishTagController")
+        leftButton = UIBarButtonItem(title: "Cancel", style: .done, target: self, action: #selector(RRTagController.cancelTagController))
+        rigthButton = UIBarButtonItem(title: "OK", style: .done, target: self, action: #selector(RRTagController.finishTagController))
         
         totalTagsSelected = 0
         self.view.addSubview(collectionTag)
+        addTagView.alpha = 0
         self.view.addSubview(addTagView)
         self.view.addSubview(controlPanelEdition)
+        navigationBar.translatesAutoresizingMaskIntoConstraints = false
         self.view.addSubview(navigationBar)
-        NSNotificationCenter.defaultCenter().addObserver(self, selector: Selector("keyboardWillShow:"), name:UIKeyboardWillShowNotification, object: nil)
-        NSNotificationCenter.defaultCenter().addObserver(self, selector: Selector("keyboardWillHide:"), name:UIKeyboardWillHideNotification, object: nil)
+        if #available(iOS 11, *) {
+            let guide = view.safeAreaLayoutGuide
+            navigationBar.prefersLargeTitles = UIDevice().userInterfaceIdiom == .phone && UIScreen.main.nativeBounds.height == 2436
+            NSLayoutConstraint.activate([
+                navigationBar.topAnchor.constraintEqualToSystemSpacingBelow(topLayoutGuide.topAnchor, multiplier: 1),
+                navigationBar.centerXAnchor.constraintEqualToSystemSpacingAfter(guide.centerXAnchor, multiplier: 1),
+                navigationBar.trailingAnchor.constraintEqualToSystemSpacingAfter(guide.trailingAnchor, multiplier: 1),
+                navigationBar.leadingAnchor.constraintEqualToSystemSpacingAfter(guide.leadingAnchor, multiplier: 1),
+                navigationBar.heightAnchor.constraint(equalToConstant: 84)
+                ])
+
+        } else {
+            NSLayoutConstraint.activate([
+                navigationBar.topAnchor.constraint(equalTo: topLayoutGuide.topAnchor),
+                navigationBar.centerXAnchor.constraint(equalTo: view.centerXAnchor),
+                navigationBar.trailingAnchor.constraint(equalTo: view.trailingAnchor),
+                navigationBar.leadingAnchor.constraint(equalTo: view.leadingAnchor),
+                navigationBar.heightAnchor.constraint(equalToConstant: 64)
+                ])
+        }
+        NotificationCenter.default.addObserver(self, selector: #selector(RRTagController.keyboardWillShow(_:)), name:NSNotification.Name.UIKeyboardWillShow, object: nil)
+        NotificationCenter.default.addObserver(self, selector: #selector(RRTagController.keyboardWillHide(_:)), name:NSNotification.Name.UIKeyboardWillHide, object: nil)
     }
     
-    class func displayTagController(#parentController: UIViewController, tagsString: [String]?,
-        blockFinish: (selectedTags: Array<Tag>, unSelectedTags: Array<Tag>)->(), blockCancel: ()->()) {
+    class func displayTagController(_ parentController: UIViewController, _ tagsString: [String]?,
+        blockFinish: @escaping (_ selectedTags: [Tag], _ unSelectedTags: [Tag]) -> Void, blockCancel: @escaping () -> Void) {
         let tagController = RRTagController()
-            tagController.tags = Array()
-            if tagsString != nil {
-                for currentTag in tagsString! {
-                    tagController.tags.append(Tag(isSelected: false, isLocked: false, textContent: currentTag))
-                }
+        tagController.tags = []
+        if tagsString != nil {
+            for currentTag in tagsString! {
+                tagController.tags.append(Tag(isSelected: false, isLocked: false, textContent: currentTag))
             }
-            tagController.blockCancel = blockCancel
-            tagController.blockFinih = blockFinish
-            parentController.presentViewController(tagController, animated: true, completion: nil)
+        }
+        tagController.blockCancel = blockCancel
+        tagController.blockFinish = blockFinish
+        parentController.present(tagController, animated: true, completion: nil)
     }
 
-    class func displayTagController(#parentController: UIViewController, tags: [Tag]?,
-        blockFinish: (selectedTags: Array<Tag>, unSelectedTags: Array<Tag>)->(), blockCancel: ()->()) {
-            let tagController = RRTagController()
-            tagController.tags = tags
-            tagController.blockCancel = blockCancel
-            tagController.blockFinih = blockFinish
-            parentController.presentViewController(tagController, animated: true, completion: nil)
+    class func displayTagController(_ parentController: UIViewController, _ tags: [Tag]?,
+        blockFinish: @escaping (_ selectedTags: [Tag], _ unSelectedTags: [Tag]) -> Void, blockCancel: @escaping () -> Void) {
+        let tagController = RRTagController()
+        tagController.tags = tags ?? []
+        tagController.blockCancel = blockCancel
+        tagController.blockFinish = blockFinish
+        parentController.present(tagController, animated: true, completion: nil)
     }
     
 }
